% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{Hann}
\alias{Hann}
\title{Hann window filter.}
\usage{
Hann(raw_data, buffer_size = 5)
}
\arguments{
\item{raw_data}{Data upon which the algorithm is applied}

\item{buffer_size}{number of points the algorithm use to compute the coefficients of the Hann window}
}
\value{
Smoothed data using Hann Window filter
}
\description{
This function return the data smoothed using the an Hann window filter.
Data are smoothed using a cosine window.
}
\examples{
raw_data = c(1:100)
smoothed_data = Hann(raw_data)

}
