% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{MA}
\alias{MA}
\title{Moving average filter.}
\usage{
MA(raw_data, buffer_size = 5)
}
\arguments{
\item{raw_data}{Data upon which the algorithm is applied}

\item{buffer_size}{number of points the algorithm use to compute the average}
}
\value{
Smoothed data using moving average algorithm
}
\description{
This function return the data smoothed using the basic moving average
algorithm. For each chunk of data of size equal to the buffer_size parameter
is calculated the average and this value is used as the i term of the newly
smoothed data.
zero padding is applied for initial and final values
}
\examples{
raw_data = c(1:100)
smoothed_data = MA(raw_data)

}
