% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{RMS}
\alias{RMS}
\title{Running median smoothing.}
\usage{
RMS(raw_data, buffer_size = 5)
}
\arguments{
\item{raw_data}{Data upon which the algorithm is applied}

\item{buffer_size}{number of points the algorithm use to compute the median}
}
\value{
Smoothed data using running median algorithm
}
\description{
This function return the data smoothed using the running median
algorithm. For each chunk of data of size equal to the buffer_size parameter
is calculated the median and this value is used as the i term of the newly
smoothed data.
For initial and final values zero padding is applied.
}
\examples{
raw_data = c(1:100)
smoothed_data = RMS(raw_data)

}
