% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{sine}
\alias{sine}
\title{Sine window filter.}
\usage{
sine(raw_data, buffer_size = 5)
}
\arguments{
\item{raw_data}{Data upon which the algorithm is applied}

\item{buffer_size}{number of points the algorithm use to compute the coefficients of the Hann window}
}
\value{
Smoothed data using Hann Window filter
}
\description{
This function return the data smoothed using the a sine window filter.
}
\examples{
raw_data = c(1:100)
smoothed_data = sine(raw_data)

}
